Added the `Sanford::Protocol` mixin
This file is the literal spec for Sanford's communication protocol and is
separated to help with updating the protocol version. If anything in the file
changes, the protocol version should probably be modified. This module can be
mixed in to make use of it's helper methods, or they can be called directly if
desired.