Added the Connection class
The Connection class takes a socket and provides a `read` and `write` method for
easily reading and writing Sanford messages. The Connection class is generic and
doesn't know if it is reading/writing a request or response. I also provided a
fake socket class that can be passed to a Connection for testing.