Protocol and connection handling redesign
This refactors how the protocol is handled and how the connection
uses the protocol classes to read and write messages.

The protocol is modeled with a handler each for the size, version,
and body portions of a message.  The size and version portions
are modeled using a `PackedHeader`.  The body portion is modeled
using a `BsonBody`.  Each handler provides a `serialize` and
`deserialize` method to encapsulate the corresponding logic.

The `Connection` is constructed given a tcp socket.  It wraps the
tcp socket with its own `Socket` handler.  This handler mixes in
`Sanford::Protocol` and encapsulates the logic of reading and
writing data from the tcp socket.  The connection then provides
a `read` and a `write` methods that take and return hash data.

Each message portion read from the socket is modeled as an
`IncomingValue`.  These helper classes handle the business logic
and error handling specific to reading each portion of the message.