Rework the FakeSocket API and update the corresponding helpers
There are API changes. First, you add incoming data to the socket
using the constructor. To read incoming or outgoing data, use the
in and out readers. The socket reader has been removed.

I added a whole slew of tests for both the fake socket and the
helpers. I also added some docs to the README.

Closes #6.