Modify requests to raise an exception when they are invalid
This is to setup stringifying the hash keys for request params. The problem is
if we don't verify that it's a hash before trying to stringify the keys, it will
raise errors. I think the better solution for moving forward is to raise a
`BadRequestError`, which effectively would ensure params are a `Hash` before
moving forward.