Add `close_write` to `Connection`
This adds a `close_write` method to the connection class. This
will allow the Sanford server and AndSon client to close the
write stream of the socket after they have written a request or
response. The benefit is this alerts the remote socket that we
are done writing and they can begin reading.