Remove version from a `Request`
This removes having a separate version string for every request.
This goes to the more generic name and params only. For versioned
services, the version should be included as part of the service's
name. This is more flexible and in Sanford, make it more clear
that services can be versioned independently.