Add `StringifyParams` module for normalizing request params
This adds a `StringifyParams` module that is used to normalize
request params. This will be used by Sanford and AndSon to
serialize params in their test helpers. This is to help ensure
that params used in tests are valid and help make users tests
more robust.

The `StringifyParams` logic was pulled out of the `Request` class.
Building a request would stringify its params when converted to a
hash.

This makes one functionality change. The `Request` now raises an
`InvalidError` which is under the `Request` namespace. This
matches our latest conventions and what was done in Qs. The error
is still raised in the same scenarios though.

This also cleans up the tests for the `Sanford::Protocol` module
and the `Request` class. This changes any static data to use
randomized data from the factory. It also updates them to our
latest conventions. This involved essentially reworking the request
tests.