response: rework status/handling
This is a number of small reworks to the response status, how it is
specified and how it is handled.  The goal here is to simplify the
logic a bit based on known usage patterns.  However, there are some
backwards incompatible changes here.

The need for simplifications was originally brought to attention
when testing this gem in modern ruby versions.  In 1.8.7, you lookup
a Hash value's key with the `index` method.  In modern ruby, you
look it up using the `key` method (ugh).  The previous status code
and name lookup logic used `index` to lookup a key if a value was
given.  To support this old logic in modern ruby, I'd have to add
a 1.8.7 conditional check (double-ugh).  I really don't like having
to have version checks in the code (although ruby is kinda forcing
it on this issue b/c they won't alias 1.8.7 methods) so Collin and
I decided to just simplify the feature.

This changes so that you can only specify response statuses using
a code integer (where previously you could use either the code
integer or a designated name string).  Dropping lookup by name
string allowed reworking the names lookup to be by code integer and
removed the need for the index/key method.  This is also simpler
overall (we found we never used the dumb name strings in practice
anyway) and also allows cleaning up the names to not use underscores
(which is just a minor aesthetic improvement).

In addition, there are a few other minor tweaks happening:

* added a formal named status for 422 ("INVALID")
* updates the response-related tests to get them up to our latest
  conventions
* switched to using randomized factory data where possible
* more thoroughly tested the response status
* tested each code/name mapping explicitly
* a bunch of minor style cleanups